image-resizer/requirements.txt
# Requirements for PNG image resizer
Pillow>=9.0.0
